+++ spec/app/models/augmentation.rb
@@ -1,0 +1,11 @@
+class Augmentation
+  include Mongoid::Document
+
+  field :name
+
+  embedded_in :player, :inverse_of => :augmentation
+
+  after_build do
+    self.name = "Infolink (#{player.frags})"
+  end
+end